refactor out ConstanceForm and get_values into forms.py and utils.py respectively
