+++ constance/admin.py
@@ -1,184 +1,23 @@
 from collections import OrderedDict
-from datetime import datetime, date, time, timedelta
-from decimal import Decimal
+from datetime import date, datetime
 from operator import itemgetter
-import hashlib
 
-from django import forms, VERSION, conf
+from django import VERSION, forms
 from django.apps import apps
 from django.contrib import admin, messages
-from django.contrib.admin import widgets
 from django.contrib.admin.options import csrf_protect_m
-from django.core.exceptions import PermissionDenied, ImproperlyConfigured
-from django.core.files.storage import default_storage
-from django.forms import fields
+from django.core.exceptions import PermissionDenied
 from django.http import HttpResponseRedirect
 from django.template.response import TemplateResponse
-from django.utils import timezone
-from django.utils.encoding import smart_bytes
-from django.utils.formats import localize
-from django.utils.module_loading import import_string
-from django.utils.text import normalize_newlines
-from django.utils.translation import gettext_lazy as _
 from django.urls import path
+from django.utils.formats import localize
+from django.utils.translation import gettext_lazy as _
 
 from . import LazyConfig, settings
-from .checks import get_inconsistent_fieldnames
-
+from .forms import ConstanceForm
+from .utils import get_values
 
 config = LazyConfig()
-
-
-NUMERIC_WIDGET = forms.TextInput(attrs={'size': 10})
-
-INTEGER_LIKE = (fields.IntegerField, {'widget': NUMERIC_WIDGET})
-STRING_LIKE = (fields.CharField, {
-    'widget': forms.Textarea(attrs={'rows': 3}),
-    'required': False,
-})
-
-FIELDS = {
-    bool: (fields.BooleanField, {'required': False}),
-    int: INTEGER_LIKE,
-    Decimal: (fields.DecimalField, {'widget': NUMERIC_WIDGET}),
-    str: STRING_LIKE,
-    datetime: (
-        fields.SplitDateTimeField, {'widget': widgets.AdminSplitDateTime}
-    ),
-    timedelta: (
-        fields.DurationField, {'widget': widgets.AdminTextInputWidget}
-    ),
-    date: (fields.DateField, {'widget': widgets.AdminDateWidget}),
-    time: (fields.TimeField, {'widget': widgets.AdminTimeWidget}),
-    float: (fields.FloatField, {'widget': NUMERIC_WIDGET}),
-}
-
-
-def parse_additional_fields(fields):
-    for key in fields:
-        field = list(fields[key])
-
-        if len(field) == 1:
-            field.append({})
-
-        field[0] = import_string(field[0])
-
-        if 'widget' in field[1]:
-            klass = import_string(field[1]['widget'])
-            field[1]['widget'] = klass(
-                **(field[1].get('widget_kwargs', {}) or {})
-            )
-
-            if 'widget_kwargs' in field[1]:
-                del field[1]['widget_kwargs']
-
-        fields[key] = field
-
-    return fields
-
-
-FIELDS.update(parse_additional_fields(settings.ADDITIONAL_FIELDS))
-
-
-def get_values():
-    """
-    Get dictionary of values from the backend
-    :return:
-    """
-
-    # First load a mapping between config name and default value
-    default_initial = ((name, options[0])
-                       for name, options in settings.CONFIG.items())
-    # Then update the mapping with actually values from the backend
-    initial = dict(default_initial, **dict(config._backend.mget(settings.CONFIG)))
-
-    return initial
-
-
-class ConstanceForm(forms.Form):
-    version = forms.CharField(widget=forms.HiddenInput)
-
-    def __init__(self, initial, request=None, *args, **kwargs):
-        super().__init__(*args, initial=initial, **kwargs)
-        version_hash = hashlib.sha256()
-
-        only_view = request and not request.user.has_perm('constance.change_config')
-        if only_view:
-            messages.warning(
-                request,
-                _("You don't have permission to change these values"),
-            )
-
-        for name, options in settings.CONFIG.items():
-            default = options[0]
-            if len(options) == 3:
-                config_type = options[2]
-                if config_type not in settings.ADDITIONAL_FIELDS and not isinstance(default, config_type):
-                    raise ImproperlyConfigured(_("Default value type must be "
-                                                 "equal to declared config "
-                                                 "parameter type. Please fix "
-                                                 "the default value of "
-                                                 "'%(name)s'.")
-                                               % {'name': name})
-            else:
-                config_type = type(default)
-
-            if config_type not in FIELDS:
-                raise ImproperlyConfigured(_("Constance doesn't support "
-                                             "config values of the type "
-                                             "%(config_type)s. Please fix "
-                                             "the value of '%(name)s'.")
-                                           % {'config_type': config_type,
-                                              'name': name})
-            field_class, kwargs = FIELDS[config_type]
-            if only_view:
-                kwargs['disabled'] = True
-            self.fields[name] = field_class(label=name, **kwargs)
-
-            version_hash.update(smart_bytes(initial.get(name, '')))
-        self.initial['version'] = version_hash.hexdigest()
-
-    def save(self):
-        for file_field in self.files:
-            file = self.cleaned_data[file_field]
-            self.cleaned_data[file_field] = default_storage.save(file.name, file)
-
-        for name in settings.CONFIG:
-            current = getattr(config, name)
-            new = self.cleaned_data[name]
-
-            if isinstance(new, str):
-                new = normalize_newlines(new)
-
-            if conf.settings.USE_TZ and isinstance(current, datetime) and not timezone.is_aware(current):
-                current = timezone.make_aware(current)
-
-            if current != new:
-                setattr(config, name, new)
-
-    def clean_version(self):
-        value = self.cleaned_data['version']
-
-        if settings.IGNORE_ADMIN_VERSION_CHECK:
-            return value
-
-        if value != self.initial['version']:
-            raise forms.ValidationError(_('The settings have been modified '
-                                          'by someone else. Please reload the '
-                                          'form and resubmit your changes.'))
-        return value
-
-    def clean(self):
-        cleaned_data = super().clean()
-
-        if not settings.CONFIG_FIELDSETS:
-            return cleaned_data
-
-        if get_inconsistent_fieldnames():
-            raise forms.ValidationError(_('CONSTANCE_CONFIG_FIELDSETS is missing '
-                                          'field(s) that exists in CONSTANCE_CONFIG.'))
-
-        return cleaned_data
 
 
 class ConstanceAdmin(admin.ModelAdmin):
